move stuff in src directory
